backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Pillow==11.0.0
anthropic==0.40.0
python-dotenv==1.0.1
requests==2.32.3
